feat(api): add pagination support for Cidade and Pessoa endpoints and improved repository

+++ src/main/java/com/vsm/gerenciador_pessoa_cidade/dtos/PessoaDTO.java
@@ -13,7 +13,7 @@ public class PessoaDTO {
     private String numero;
     private String bairro;
     private String cep;
-    private int cidadeId;
+    private Long cidadeId;
     private String telefone;
     private String email;
 }
